Float64 dtype was returned. Stifel manifold is ready (probably).

+++ src/spaces/hyperbolic.py
@@ -7,7 +7,7 @@ from math import sqrt
 device = 'cuda' if torch.cuda.is_available() else 'cpu'
 #device = 'cpu'
 
-dtype = torch.float32
+dtype = torch.float64
 j = torch.tensor([1j], device=device).item()  # imaginary unit
 pi = 2*torch.acos(torch.zeros(1)).item()
